made randomizeAttachments into a helper function to clean up the attachment apis
