+++ pages/api/black-ops-six/attachments/sniper.tsx
@@ -1,7 +1,7 @@
 import lr762 from "@/json/black-ops-six/attachments/sniper/lr762.json";
 import lw3a1Frostline from "@/json/black-ops-six/attachments/sniper/lw3a1Frostline.json";
 import svd from "@/json/black-ops-six/attachments/sniper/svd.json";
-import { randomListItem } from "@/helpers/randomListItem";
+import { randomizeAttachments } from "@/helpers/randomizeAttachments";
 
 export default async function handler(req, res) {
   let attachments: any = {};
@@ -27,33 +27,5 @@ function getGunAttachments(gun: string) {
       return svd;
     default:
       return null;
-  }
-}
-
-/**
- * Randomly selects attachments from a pool of data.
- *
- * @param {object} attachArr - The object to store the selected attachments.
- * @param {object} data - The pool of attachments to choose from.
- * @param {number} count - The number of attachments to select.
- *
- * @returns {void}
- */
-function randomizeAttachments(attachArr: any, data: any, count: number) {
-  let attachs = 0;
-
-  while (attachs < count) {
-    const keys = Object.keys(data);
-    const randomIndex = Math.floor(Math.random() * keys.length);
-    const randomKey = keys[randomIndex];
-
-    if (attachArr.hasOwnProperty(randomKey)) {
-      console.log("randomizeAttachments: already exists");
-      continue;
-    } else {
-      console.log("randomizeAttachments: doesnt already exist");
-      attachArr[randomKey] = randomListItem(data[randomKey]);
-      attachs++;
-    }
   }
 }
